requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
python-telegram-bot==13.15
pytz==2024.1
requests==2.31.0
six==1.16.0
tornado==6.1
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
